requirements.txt
asgiref==3.2.9
backcall==0.2.0
decorator==4.4.2
Django==3.0.7
djangorestframework==3.11.0
gunicorn==20.0.4
ipython==7.15.0
ipython-genutils==0.2.0
jedi==0.17.0
parso==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
pkg-resources==0.0.0
prompt-toolkit==3.0.5
ptyprocess==0.6.0
Pygments==2.6.1
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
traitlets==4.3.3
wcwidth==0.2.4
